rename: 정적 리소스 관련 Handler 패키지 변경
