+++ tomcat/src/main/java/mvc/controller/LoginController.java
@@ -1,6 +1,6 @@
 package mvc.controller;
 
-import nextstep.handler.exception.InvalidQueryParameterException;
+import mvc.controller.exception.InvalidQueryParameterException;
 import nextstep.jwp.application.UserService;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http.HttpCookie;
