+++ tomcat/src/main/java/mvc/controller/exception/InvalidQueryParameterException.java
@@ -1,4 +1,4 @@
-package nextstep.handler.exception;
+package mvc.controller.exception;
 
 public class InvalidQueryParameterException extends IllegalArgumentException {
