+++ tomcat/src/main/java/nextstep/jwp/application/exception/LoginFailureException.java
@@ -1,4 +1,4 @@
-package nextstep.handler.exception;
+package nextstep.jwp.application.exception;
 
 public class LoginFailureException extends IllegalArgumentException {
