+++ tomcat/src/main/java/org/apache/coyote/handler/ResourceHandler.java
@@ -1,7 +1,7 @@
-package nextstep.handler;
+package org.apache.coyote.handler;
 
 import java.io.IOException;
-import nextstep.handler.util.ResourceProcessor;
+import org.apache.coyote.handler.util.ResourceProcessor;
 import org.apache.coyote.Handler;
 import org.apache.coyote.http.request.Request;
 import org.apache.coyote.http.response.ContentType;
